bugfix: update unique_logical_size from long to double

+++ cloudtower/models/iscsi_lun.py
@@ -61,7 +61,7 @@ class IscsiLun(object):
         'stripe_num': 'int',
         'stripe_size': 'int',
         'thin_provision': 'bool',
-        'unique_logical_size': 'int',
+        'unique_logical_size': 'float',
         'unique_size': 'int',
         'zbs_volume_id': 'str'
     }
@@ -1056,7 +1056,7 @@ class IscsiLun(object):
 
 
         :return: The unique_logical_size of this IscsiLun.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._unique_logical_size
 
@@ -1066,7 +1066,7 @@ class IscsiLun(object):
 
 
         :param unique_logical_size: The unique_logical_size of this IscsiLun.  # noqa: E501
-        :type unique_logical_size: int
+        :type unique_logical_size: float
         """
 
         self._unique_logical_size = unique_logical_size
